An analysis of interesting data related to Covid
With the use of sqlite3 and pandas and numpy libraries, the following project aims at finding interesting correlations that happened around the Covid-19 pandemic. The logical and physical design of an SQL database has also been created. Consult to the code and the file named "ER diagram" for more information.